feat: Create get_pretrained_segment_borzoi_model
